Implementa a funcionalidade adicionar ao carrinho
