+++ package.json
@@ -23,9 +23,9 @@
     "gulp-babel": "^8.0.0",
     "gulp-clean-css": "^4.2.0",
     "gulp-concat": "^2.6.1",
-    "gulp-connect": "^5.7.0",
     "gulp-imagemin": "^6.1.0",
     "gulp-open": "^3.0.1",
-    "gulp-uglify": "^3.0.2"
+    "gulp-uglify": "^3.0.2",
+    "json-server": "^0.15.0"
   }
 }
