+++ src/styles/responsive.css
@@ -1,3 +1,5 @@
+/* Estilização responsiva */
+
 /* media queries */
 @media(max-width: 480px) {
     header {
@@ -37,6 +39,12 @@
         grid-column: 1 / 2;
         justify-self: center;
     } 
+
+    main p#products-not-found {
+        grid-column: 1 / 2;
+        justify-self: center;
+        font-size: 1.3rem;
+    }
 }
 
 @media(min-width: 481px) {
@@ -73,6 +81,11 @@
     }
 
     main button#load-more-products {
+        grid-column: 1 / 3;
+        justify-self: center;
+    }
+
+    main p#products-not-found {
         grid-column: 1 / 3;
         justify-self: center;
     }
@@ -147,6 +160,11 @@
 
     main button#load-more-products {
         grid-column: 2 / 3;
+        justify-self: end;
+    }
+
+    main p#products-not-found {
+        grid-column: 1 / 3;
         justify-self: end;
     }
 }
